2544 Fixed opening message not appearing
The opening message was not appearing when a user would pick up a new chat, this is because the input area was not cleared before the check to see if the it was a new chat was made. To resolve, when a new chat is picked up (not monitored). The input area is cleared first, allowing the opening message to be added.